backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
pytz==2023.3.post1
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
psycopg2==2.9.6
whitenoise==6.5.0
Pillow==10.0.1